Moved and renamed record dictionary and taxonomy files
